+++ ncmb-core/build.gradle
@@ -4,10 +4,10 @@ buildscript {
     repositories {
         google()
         jcenter()
+        google()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.0.0'
-
     }
 }
 
@@ -39,29 +39,28 @@ android {
     }
 }
 repositories {
-    maven {
-        url 'https://maven.google.com'
-        // Alternative URL is 'https://dl.google.com/dl/android/maven2/'
-    }
     google()
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.google.android.gms:play-services-gcm:11.0.0'
+    //noinspection GradleCompatible
+    implementation 'com.google.firebase:firebase-messaging:15.0.0'
+    implementation 'com.google.android.gms:play-services-base:15.0.0'
     implementation 'com.google.code.gson:gson:2.3.1'
-    implementation 'com.android.support:appcompat-v7:26.0.2'
+    implementation 'com.android.support:appcompat-v7:26.1.0'
+
     testImplementation 'junit:junit:4.12'
-    testImplementation 'org.robolectric:robolectric:3.0'
+    testImplementation "org.robolectric:robolectric:3.0"
     testImplementation 'org.robolectric:shadows-httpclient:3.0'
     testImplementation 'com.squareup.okhttp:mockwebserver:2.3.0'
-    testImplementation 'org.mockito:mockito-core:1.10.19'
+    testImplementation 'org.mockito:mockito-core:1.+'
     testImplementation 'org.powermock:powermock-api-mockito:1.6.5'
     testImplementation 'org.powermock:powermock-module-junit4:1.6.5'
     testImplementation 'org.powermock:powermock-module-junit4-rule:1.6.5'
     testImplementation 'org.powermock:powermock-classloading-xstream:1.6.5'
     testImplementation 'org.yaml:snakeyaml:1.15'
-    testImplementation 'org.skyscreamer:jsonassert:1.2.3'
+    testImplementation 'org.skyscreamer:jsonassert:1.2.+'
     testImplementation "org.robolectric:shadows-play-services:3.0"
     testImplementation "org.robolectric:shadows-support-v4:3.0"
 }
