+++ ncmb-core/src/test/assets/yaml/mbaas.yml
